# protocol/demo.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "demo.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "DemoService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/demo": {
      "post": {
        "summary": "单次调用的方法 支持http和grpc调用",
        "operationId": "DemoService_OneWay",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/exampleRespPkg"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/exampleReqPkg"
            }
          }
        ],
        "tags": [
          "DemoService"
        ]
      }
    },
    "/halfstream": {
      "post": {
        "summary": "流式调用的方法 支持http chunked和grpc调用",
        "operationId": "DemoService_HalfStream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/exampleRespPkg"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of exampleRespPkg"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/exampleReqPkg"
            }
          }
        ],
        "tags": [
          "DemoService"
        ]
      }
    },
    "/stream": {
      "post": {
        "summary": "流式调用的方法 支持http chunked和grpc调用",
        "operationId": "DemoService_Stream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/exampleRespPkg"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of exampleRespPkg"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/exampleReqPkg"
            }
          }
        ],
        "tags": [
          "DemoService"
        ]
      }
    }
  },
  "definitions": {
    "exampleReqPkg": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "/ 用户名"
        },
        "age": {
          "type": "integer",
          "format": "int32",
          "title": "/ 名字"
        }
      },
      "title": "*\r\n响应包"
    },
    "exampleRespPkg": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "title": "/ 响应码"
        },
        "msg": {
          "type": "string",
          "title": "/ 响应消息"
        }
      },
      "title": "*\r\n响应包"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}